Implement loading data from JSON files

+++ shared/src/commonMain/kotlin/org/dishch/medcalculator/data/local/DosageRegimenEntity.kt
@@ -22,7 +22,7 @@ data class DosageRegimenEntity(
     @PrimaryKey(autoGenerate = true) val id: Long = 0,
     @ColumnInfo(name = "from_age") val fromAge: Int,
     @ColumnInfo(name = "to_age") val toAge: Int,
-    @ColumnInfo(name = "min_dose_per_kg") val minDosePerKg: Int,
-    @ColumnInfo(name = "max_dose_per_kg") val maxDosePerKg: Int,
+    @ColumnInfo(name = "min_dose_per_kg") val minDosePerKg: Double,
+    @ColumnInfo(name = "max_dose_per_kg") val maxDosePerKg: Double,
     @ColumnInfo(name = "medication_id") val medicationId: Long
 )
